# Refactor superblock pack/unpack