--- a/src/Makefile.test.include
+++ b/src/Makefile.test.include
@@ -16,6 +16,11 @@ JSON_TEST_FILES = \
   test/data/tx_invalid.json.h \
   test/data/tx_valid.json.h
 
+OTHER_TEST_FILES = \
+  test/data/superblock.txt.h \
+  test/data/superblock_packed.bin.h \
+  test/data/superblock_unpacked.txt.h
+
 # test_n binary #
 GRIDCOIN_TESTS =\
   test/appcache_tests.cpp \
@@ -48,7 +53,7 @@ GRIDCOIN_TESTS =\
   test/wallet_tests.cpp
 
 
-test_test_gridcoin_SOURCES = $(GRIDCOIN_TESTS) $(JSON_TEST_FILES)
+test_test_gridcoin_SOURCES = $(GRIDCOIN_TESTS) $(JSON_TEST_FILES) $(OTHER_TEST_FILES)
 test_test_gridcoin_CPPFLAGS = $(AM_CPPFLAGS) $(GRIDCOIN_INCLUDES) -I$(builddir)/test/ $(TESTDEFS) $(EVENT_CFLAGS)
 test_test_gridcoin_LDADD = $(LIBGRIDCOIN_UTIL) $(LIBUNIVALUE) $(LIBLEVELDB) $(LIBMEMENV) $(BOOST_LIBS) $(BOOST_UNIT_TEST_FRAMEWORK_LIB) $(EVENT_LIBS) $(EVENT_PTHREADS_LIBS)
 test_test_gridcoin_CXXFLAGS = $(AM_CXXFLAGS) $(PIE_FLAGS)