--- a/src/contract/superblock.cpp
+++ b/src/contract/superblock.cpp
@@ -22,45 +22,43 @@ std::string UnpackBinarySuperblock(std::string sBlock)
     std::string sBinary = ExtractXML(sBlock,"<BINARY>","</BINARY>");
     if (sBinary.empty()) return sBlock;
 
+    std::ostringstream stream;
+    stream << "<AVERAGES>" << ExtractXML(sBlock,"<AVERAGES>","</AVERAGES>") << "</AVERAGES>"
+           << "<QUOTES>" << ExtractXML(sBlock,"<QUOTES>","</QUOTES>") << "</QUOTES>"
+           << "<MAGNITUDES>";
+
     // Binary data support structure:
     // Each CPID consumes 16 bytes and 2 bytes for magnitude: (Except CPIDs with zero magnitude - the count of those is stored in XML node <ZERO> to save space)
     // 1234567890123456MM
     // MM = Magnitude stored as 2 bytes
     // No delimiter between CPIDs, Step Rate = 18.
     // CPID and magnitude are stored in big endian.
-    std::string sReconstructedMagnitudes;
     for (unsigned int x = 0; x < sBinary.length(); x += 18)
     {
         if(sBinary.length() - x < 18)
             break;
 
         const BinaryResearcher* researcher = reinterpret_cast<const BinaryResearcher*>(sBinary.data() + x);
-        sReconstructedMagnitudes +=
-                HexStr(researcher->cpid.begin(), researcher->cpid.end()) +
-                "," +
-                ToString(researcher->magnitude) + ";";
+        stream << HexStr(researcher->cpid.begin(), researcher->cpid.end()) << ","
+               << researcher->magnitude << ";";
     }
 
     // Append zero magnitude researchers so the beacon count matches
     int num_zero_mag = atoi(ExtractXML(sBlock,"<ZERO>","</ZERO>"));
     const std::string zero_entry("0,15;");
-    sReconstructedMagnitudes.reserve(zero_entry.size() * num_zero_mag);
     for(int i=0; i<num_zero_mag; ++i)
-        sReconstructedMagnitudes += zero_entry;
+        stream << zero_entry;
 
-    std::string sAverages   = ExtractXML(sBlock,"<AVERAGES>","</AVERAGES>");
-    std::string sQuotes     = ExtractXML(sBlock,"<QUOTES>","</QUOTES>");
-    return "<AVERAGES>" + sAverages + "</AVERAGES><QUOTES>" + sQuotes + "</QUOTES><MAGNITUDES>" + sReconstructedMagnitudes + "</MAGNITUDES>";
+    stream << "</MAGNITUDES>";
+    return stream.str();
 }
 
 std::string PackBinarySuperblock(std::string sBlock)
 {
     std::string sMagnitudes = ExtractXML(sBlock,"<MAGNITUDES>","</MAGNITUDES>");
-    std::string sAverages   = ExtractXML(sBlock,"<AVERAGES>","</AVERAGES>");
-    std::string sQuotes     = ExtractXML(sBlock,"<QUOTES>","</QUOTES>");
 
     // For each CPID in the superblock, convert data to binary
-    std::string sBinary = "";
+    std::stringstream stream;
     int64_t num_zero_mag = 0;
     for (auto& entry : split(sMagnitudes.c_str(), ";"))
     {
@@ -82,9 +80,13 @@ std::string PackBinarySuperblock(std::string sBlock)
         magnitude_d = std::max(0.0, std::min(magnitude_d, 32767.0));
         researcher.magnitude = roundint(magnitude_d);
 
-        sBinary.append((unsigned char*) &researcher,
-                       (unsigned char*) (&researcher) + sizeof(BinaryResearcher));
+        stream.write((const char*) &researcher, sizeof(BinaryResearcher));
     }
 
-    return "<ZERO>" + i64tostr(num_zero_mag) + "</ZERO><BINARY>" + sBinary + "</BINARY><AVERAGES>" + sAverages + "</AVERAGES><QUOTES>" + sQuotes + "</QUOTES>";
+    std::stringstream block_stream;
+    block_stream << "<ZERO>" << num_zero_mag << "</ZERO>"
+                    "<BINARY>" << stream.rdbuf() << "</BINARY>"
+                    "<AVERAGES>" << ExtractXML(sBlock,"<AVERAGES>","</AVERAGES>") << "</AVERAGES>"
+                    "<QUOTES>" << ExtractXML(sBlock,"<QUOTES>","</QUOTES>") << "</QUOTES>";
+    return block_stream.str();
 }