--- a/src/contract/superblock.h
+++ b/src/contract/superblock.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <string>
+
+std::string UnpackBinarySuperblock(std::string block);
+std::string PackBinarySuperblock(std::string sBlock);