**State: Ready**

Try to simplify superblock packing and unpacking to move away from the custom binary serialization code which was extremely slow.

TODO
====
- [x] Either remove boost dependency (replace with ntohs/htons) or bump the boost requirements.
 
Benchmark
========

Tested by measuring the time before and after running the pack/unpack in the unit tests. Times are in microseconds.

| Arch  | Before | After | Speedup |
|------|--------|------|-----------|
| i5 Pack      | 34003  | 1159  | 29.3x   |
| i5 Unpack    | 63120  | 403  | 156.6x   |
| Core2 Pack   |  63718 | 2212 | 28.8x |
| Core2 Unpack | 114080 | 621 | 183.7x |
| Pi 3 Pack         | ? | ? | ? |
| Pi 3 Unpack    | 566126 | 4416 | 128.2x |

63 milliseconds might not sound like that big of a deal, but we tally and unpack every 10 blocks so in a 500 block batch it adds up. Time saved on 500 v9 blocks:

- i5: 3.15s
- Core2: 5.7s
- RPi3: 28s

Help wanted
=========
The implementation still feels a bit messy so if anyone has any ideas on how to clean it up I'd be grateful.

I'd also like to test this before and after on a Raspbeery Pi to see what the speed increase looks like.