@denravonska 2018-07-25
This should also greatly help #768.