@denravonska 2018-08-10
Added a static assert to ensure compiler error if the magnitude is padded.